requirements.txt
pandas~=1.5.2
PyYAML~=6.0
numpy~=1.24.1
scikit-learn~=1.2.0
Pillow~=9.4.0